refactor: lift command builder from recipes page into commands/shared
